# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Transit Manager API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /whatsapp/config:
    get:
      operationId: getWhatsAppConfig
      tags: [whatsapp]
      summary: Get WhatsApp configuration (admin only)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: WhatsApp config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhatsAppConfig"
    put:
      operationId: updateWhatsAppConfig
      tags: [whatsapp]
      summary: Update WhatsApp configuration (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWhatsAppConfigRequest"
      responses:
        "200":
          description: WhatsApp config updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhatsAppConfig"
  /whatsapp/messages:
    get:
      operationId: listMessages
      tags: [whatsapp]
      summary: List WhatsApp messages
      security:
        - bearerAuth: []
      parameters:
        - name: contactPhone
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WhatsAppMessage"
    post:
      operationId: sendMessage
      tags: [whatsapp]
      summary: Send a WhatsApp message
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "200":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhatsAppMessage"
  /whatsapp/webhook:
    post:
      operationId: webhookReceive
      tags: [whatsapp]
      summary: Receive WhatsApp webhook events
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Webhook processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
    get:
      operationId: webhookVerify
      tags: [whatsapp]
      summary: Verify WhatsApp webhook
      parameters:
        - name: hub.mode
          in: query
          schema:
            type: string
        - name: hub.challenge
          in: query
          schema:
            type: string
        - name: hub.verify_token
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Webhook verified
          content:
            text/plain:
              schema:
                type: string
  /whatsapp/contacts:
    get:
      operationId: listContacts
      tags: [whatsapp]
      summary: List WhatsApp contacts (unique phone numbers that have messaged)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WhatsAppContact"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    WhatsAppConfig:
      type: object
      properties:
        phoneNumberId:
          type: string
          nullable: true
        accessToken:
          type: string
          nullable: true
        verifyToken:
          type: string
          nullable: true
        businessAccountId:
          type: string
          nullable: true
        isConfigured:
          type: boolean
      required:
        - isConfigured
    UpdateWhatsAppConfigRequest:
      type: object
      properties:
        phoneNumberId:
          type: string
        accessToken:
          type: string
        verifyToken:
          type: string
        businessAccountId:
          type: string
    WhatsAppMessage:
      type: object
      properties:
        id:
          type: integer
        messageId:
          type: string
          nullable: true
        contactPhone:
          type: string
        contactName:
          type: string
          nullable: true
        body:
          type: string
        direction:
          type: string
          enum: [inbound, outbound]
        status:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - contactPhone
        - body
        - direction
        - createdAt
    SendMessageRequest:
      type: object
      properties:
        to:
          type: string
        message:
          type: string
      required:
        - to
        - message
    WhatsAppContact:
      type: object
      properties:
        phone:
          type: string
        name:
          type: string
          nullable: true
        lastMessage:
          type: string
          nullable: true
        lastMessageAt:
          type: string
          format: date-time
          nullable: true
        unreadCount:
          type: integer
      required:
        - phone
        - unreadCount